+++ test/cmdt.js
@@ -151,6 +151,9 @@ buster.testCase('cmdt - _exec', {
       { command: 'whoami', file: 'file1.yml', exitcode: 111 }
     ];
     cmdt._exec(tests);
+    assert.equals(cmdt._execData['file1.yml'].output, 'somestdoutdatasomestderrdata');
+    assert.equals(cmdt._execData['file1.yml'].stdout, 'somestdoutdata');
+    assert.equals(cmdt._execData['file1.yml'].stderr, 'somestderrdata');
     done();
   }
 });
@@ -160,35 +163,35 @@ buster.testCase('cmdt - _testCb', {
     this.mockReporter = this.mock(reporter);
   },
   'should emit success to reporter when there is no error': function (done) {
-    this.mockReporter.expects('emit').withExactArgs('success', { exitCode: 0, file: 'somefile1.yml', output: 'someoutput' }, { exitcode: 0, output: 'someoutput' });
+    this.mockReporter.expects('emit').withExactArgs('success', { exitcode: 0, file: 'somefile1.yml', output: 'someoutput' }, { exitcode: 0, output: 'someoutput' });
 
-    var test = { file: 'somefile1.yml', exitCode: 0, output: 'someoutput' };
+    var test = { file: 'somefile1.yml', exitcode: 0, output: 'someoutput' };
 
     var cmdt = new Cmdt();
     cmdt._execData['somefile1.yml'] = {
-      exitCode: 0,
+      exitcode: 0,
       output  : 'someoutput'
     };
     cmdt._testCb(test, done)();
   },
   'should emit failure to reporter when there is an error': function (done) {
-    this.mockReporter.expects('emit').withExactArgs('failure', ['Output does not match expected regexp \'someotheroutput\''], { exitCode: 0, file: 'somefile2.yml', output: 'someotheroutput' }, { exitcode: 0, output: 'someoutput' });
+    this.mockReporter.expects('emit').withExactArgs('failure', ['Output does not match expected regexp \'someotheroutput\''], { exitcode: 0, file: 'somefile2.yml', output: 'someotheroutput' }, { exitcode: 0, output: 'someoutput' });
 
-    var test = { file: 'somefile2.yml', exitCode: 0, output: 'someotheroutput' };
+    var test = { file: 'somefile2.yml', exitcode: 0, output: 'someotheroutput' };
 
     var cmdt = new Cmdt();
     cmdt._execData['somefile2.yml'] = {
-      exitCode: 0,
+      exitcode: 0,
       output  : 'someoutput'
     };
     cmdt._testCb(test, done)();
   },
   'should pass test command execution error to callback': function (done) {
-    var test = { file: 'somefile3.yml', exitCode: 0, output: 'someotheroutput' };
+    var test = { file: 'somefile3.yml', exitcode: 0, output: 'someotheroutput' };
 
     var cmdt = new Cmdt();
     cmdt._execData['somefile3.yml'] = {
-      exitCode: 0,
+      exitcode: 0,
       output  : 'someoutput'
     };
     cmdt._testCb(test, function (err) {